requirements.txt
Django==4.1.2
django-filter==22.1
django-auto-logout==0.5.0
djangorestframework==3.14.0
Pillow==9.3.0

